+++ benchmarks/install_and_test.sh
@@ -66,10 +66,10 @@ function runNodeBenchmark(){
 }
 
 function runCSharpBenchmark(){
-  cd ${CSHARP_BENCH_FOLDER}
+  cd ${GLIDE_HOME_FOLDER}/../benchmarks
   dotnet clean
   dotnet build --configuration Release /warnaserror
-  dotnet run --framework $dotnetFramework --configuration Release --resultsFile=../$1 --dataSize $2 --concurrentTasks $concurrentTasks --clients $chosenClients --host $host --clientCount $clientCount $tlsFlag $portFlag $minimalFlag
+  dotnet run --framework $dotnetFramework --configuration Release --resultsFile=$1 --dataSize $2 --concurrentTasks $concurrentTasks --clients $chosenClients --host $host --clientCount $clientCount $tlsFlag $portFlag $minimalFlag
 }
 
 function runJavaBenchmark(){